don't use textWidth if signal name is a tspan
closes #299

We were using `getBBox().width` before "browserless version" but we
cannot do this anymore with tspans. Therefore, we skip measuring
for this lane. This will only work if there are other lanes that are
wider and have a normal name. For cases when this doesn't work, the
user must explicitely set textWidth on the tspan.